test: replace unit tests with integration tests for FeeCalculator
Following PR review feedback, replace mocked FileManager tests with
integration tests that use real file operations. This provides better
test coverage and validates actual file I/O behavior.

Tests now:
- Create and clean up temporary test directories
- Write actual test data to FileManager
- Read back and verify fee reports
- Test all edge cases with real file operations
